requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
Babel==2.2.0              # via babeldjango, django-prices
BabelDjango==0.2.2        # via django-prices
boto==2.39.0              # via django-offsite-storage
braintree==3.25.0         # via django-payments
cffi==1.5.2               # via cryptography
cryptography==1.2.3       # via django-payments
cssselect==0.9.1          # via premailer
cssutils==1.0.1           # via premailer
dj-database-url==0.4.0
dj-email-url==0.0.4
django-absolute==0.3      # via django-emailit
django-bootstrap3==7.0.0
django-cache-url==1.0.0
django-countries==3.4.1
django-emailit==0.2.2
django-materializecss-form==1.0.1
django-model-utils==2.4
django-mptt==0.8.3
django-offsite-storage==0.0.9
django-payments==0.9.2
django-prices==0.4.13
django-redis==4.3.0
django-selectable==0.9.0
django-versatileimagefield==1.3
Django==1.9.3             # via django-absolute, django-emailit, django-model-utils, django-mptt, django-offsite-storage, django-payments, django-prices, jsonfield
djangorestframework==3.3.2
fake-factory==0.5.5
google-i18n-address==1.0.7
google-measurement-protocol==0.1.5
idna==2.0                 # via cryptography
jsonfield==1.0.3
lxml==3.5.0               # via premailer
Markdown==2.6.5
maxminddb-geolite2==2015.708
maxminddb==1.2.0
mock==1.3.0
pbr==1.8.1                # via mock
Pillow==3.1.1             # via django-versatileimagefield
premailer==2.9.7          # via django-emailit
prices==0.5.8
psycopg2==2.6.1
purl==1.2
py==1.4.31                # via pytest
pyasn1==0.1.9             # via cryptography
pycparser==2.14           # via cffi
PyJWT==1.4.0              # via django-payments
pytest-django==2.9.1
pytest==2.9.0
python-dateutil==2.5.0    # via fake-factory
pytz==2015.7              # via babel
redis==2.10.5             # via django-redis
requests==2.9.1
satchless==1.1.3
six==1.10.0               # via cryptography, fake-factory, mock, purl, python-dateutil
stripe==1.30.0            # via django-payments
suds-jurko==0.6           # via django-payments
unidecode==0.4.19
uwsgi==2.0.12
xmltodict==0.10.1         # via django-payments
django-modeltranslation==0.11      # db model i18n

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via cryptography
